gl: Adds Shader class
The Shader class takes care of compiling and linking the shaders.
Currently only supports vertex and fragment shaders.

The preferred usage is to create a ShaderProgram object and initialize
it with two strings containing the vertex and fragment shader code.
Then, call the compileShaderProgram() member function to generate the
program.

Signed-off-by: Erick Narvaez <zadielerick@gmail.com>